Delete works by getting an email

+++ src/main/java/com/masifunde/masifunde/controllers/UserResource.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping(value="/rest/users")
@@ -40,9 +41,28 @@ public class UserResource {
         return userRepository.findAll();
     }
 
-    @DeleteMapping("/delete/{id}")
-    public List<User> deleteUser(@PathVariable Integer id) {
-        userRepository.deleteById(id);
+    @DeleteMapping("/delete/{email}")
+    public List<User> deleteUser(@PathVariable String email) {
+        User user = new User();
+        Integer id=0;
+
+        if(!userRepository.findByEmail(email).isEmpty()){
+            List<User> u =userRepository.findAll();
+
+            for(int i=0 ;i<u.size() ; i++)
+            {
+                String userEmail = u.get(i).getEmail();
+                System.out.println(userEmail +" >>> " + email);
+                if(email.trim().equals(userEmail))
+                {
+                    id=u.get(i).getUser_id();
+                    System.out.println(id);
+                    System.out.println(userEmail +" is the one");
+                }
+            }
+
+            userRepository.deleteById(id);
+        }
         return userRepository.findAll();
     }
 }
